added modified settings page
