+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/AppController.java
@@ -5,7 +5,6 @@ import android.app.Application;
 import android.content.Context;
 
 import com.princess.android.cryptonews.injection.AppComponent;
-import com.princess.android.cryptonews.injection.DaggerAppComponent;
 
 import javax.inject.Inject;
