+++ CryptoNewsAndroid/app/src/main/res/layout/fragment_latest_news.xml
@@ -20,4 +20,4 @@
         layout="@layout/empty_item"
         android:layout_gravity="center">
     </include>
-</FrameLayout>
+    </FrameLayout>
