+++ CryptoNewsAndroid/app/src/main/res/layout/news_list_item.xml
@@ -9,7 +9,6 @@
     android:layout_margin="@dimen/padding">
 
     <android.support.constraint.ConstraintLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_marginBottom="8dp"
         android:layout_height="match_parent">
@@ -19,7 +18,6 @@
             android:layout_width="156dp"
             android:layout_height="127dp"
             android:scaleType="centerCrop"
-            android:src="@drawable/ic_news_item"
             app:riv_corner_radius_top_left="6dp"
             app:riv_corner_radius_top_right="6dp"
             app:layout_constraintEnd_toEndOf="parent"
@@ -39,7 +37,7 @@
             android:paddingStart="6dp"
             android:paddingEnd="6dp"
             android:scrollHorizontally="false"
-            android:text="This is the latest ethereum rate cryptonews"
+            android:text=""
             android:textColor="#3E3E3E"
             android:textSize="13dp"
             app:layout_constraintEnd_toEndOf="parent"
@@ -48,14 +46,13 @@
 
 
         <TextView
-
             android:layout_marginTop="16dp"
             android:id="@+id/news_site"
             android:layout_width="wrap_content"
+            android:text=""
             android:textSize="10dp"
             android:layout_height="wrap_content"
             android:paddingStart="6dp"
-            android:text="Crypto Recorder"
             android:textColor="#27AE60"
             app:layout_constraintEnd_toStartOf="@+id/news_date"
             app:layout_constraintHorizontal_bias="0.5"
@@ -68,7 +65,7 @@
             android:layout_height="wrap_content"
             android:gravity="right"
             android:textStyle="italic"
-            android:text="1hr ago"
+            android:text=""
             android:textSize="10dp"
             android:textColor="#FF6363"
             android:paddingEnd="6dp"
@@ -78,8 +75,6 @@
             app:layout_constraintTop_toTopOf="@+id/news_site" />
 
 
-    </android.support.constraint.ConstraintLayout>
+        </android.support.constraint.ConstraintLayout>
 
-
-
-</android.support.v7.widget.CardView>
+    </android.support.v7.widget.CardView>
