+++ CryptoNewsAndroid/app/src/main/res/values/strings.xml
@@ -78,4 +78,6 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="manage_blog_websites">Manage Blog Websites</string>
+    <string name="manage_blog_title">Manage the Blog Websites you want to receive updates from. Due to scalability issues and API timeouts, you can only have a maximum of 4 websites.</string>
 </resources>
