+++ CryptoNewsAndroid/app/src/main/res/xml/pref_headers.xml
@@ -1,20 +1,9 @@
-<preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <!-- These settings headers are only used on tablets. -->
+        <Preference
+            android:defaultValue="true"
+            android:key="manage_blog"
+            android:title="@string/manage_blog_websites" />
 
-    <header
-        android:fragment="com.princess.android.cryptonews.settings.SettingsActivity$GeneralPreferenceFragment"
-        android:icon="@drawable/ic_info_black_24dp"
-        android:title="@string/pref_header_general" />
-
-    <header
-        android:fragment="com.princess.android.cryptonews.settings.SettingsActivity$NotificationPreferenceFragment"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/pref_header_notifications" />
-
-    <header
-        android:fragment="com.princess.android.cryptonews.settings.SettingsActivity$DataSyncPreferenceFragment"
-        android:icon="@drawable/ic_sync_black_24dp"
-        android:title="@string/pref_header_data_sync" />
-
-</preference-headers>
+</PreferenceScreen>
